requirements-dev.txt
-r ./requirements.txt

aioresponses==0.7.4
black==23.1.0
coverage==7.2.0
django-debug-toolbar==3.8.1
flake8==6.0.0
freezegun==1.2.2
ipdb==0.13.11
isort==5.12.0
pre-commit==3.1.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-django==4.5.2
pytest-mock==3.10.0
selenium==4.7.2
